more cleanup in Grid unit
Implementation of Grid_getBlkNeighBlkIDFromPos exists in this code,
so added it back into Grid_interface.
But dded appropriate DEV notes.

gr_findNeghID is specific to PARAMESH and must take a block ID as
first argument instead of a descriptor, since it is called like that
from an obscure place in
GridParticles/GridParticlesMove/ParticlesOwned/Paramesh/VirtualParticles/gr_ptVPGenerate.F90 .