Remove header file that claimed to be licensed under LGPL
Apart from comments, this file PM4_package/headers/paramesh_preprocessor.fh
had three lines that could be considered "code", all of them preprocessor
stuff" two includes and one define.

I do not believe the code in such a file, sans comments, can be considered
protected under copyright as an original work.

Moreover, the notice that existed in the file made claims about the
PARAMESH library clearly in contradiction with most other source files
and the text of the PARAMESH_USERS_AGREEMENT. I therefore assume
that the notice in the paramesh_preprocessor.fh fiel was left there
by mistake from an erlier version by the original authors, and was
not intended to convey additional lincense terms to what is found
elsewhere in the PARAMESH source, but was intended to be replaced
by the license terms found elsewhere.

I am removing this file here; the relevant 3 lines will be re-created
elsewhere.