workaround for installations where HDF5 Fortran calls fail
Define SUPPRESS_HDF5_FORTRAN_CALLS (maybe in a Makefile.h
specific to site and compiler) to skip Fortran code in
io_writeGrid.F90 that makes calls to HDF5 subprograms.
If io_writeGrid actually gets called, the code will abort.

This workaround is for the environment of newer GCE machines,
where I could so far not get Flash-X to work properly
with a library thart provides HDF5 Fortran bindings.

Note that the workaround does not make HDF5 completely
invisible in the Fortran code: the line

>   use HDF5

is outside of the region of code that gets suppressed
when SUPPRESS_HDF5_FORTRAN_CALLS is defined. This works
on the GCE machines in question (probably depending on
activating the right Lmode modules), but may not work in
other situations where the Flash-X dependence on HDF5
Fortran bindings needs to be disabled for some reason.