tweak bin/libUtils.py to support reuse of library build
Some tweaks in setup's bin/libUtils.py.
These are meant to allow the intended behavior
if the libinfo.py of a library under lib
(of INTERNAL type) does **not** include
"REBUILD":1 in its answer: in that case, an
already existing lib<NAME>.a file (in the right
location) should be left in place and can be
reused, instead of being built again every time
setup.py is run.

(Also typo corrections in bin/Makefile.tpl.)