Modify libinfo.py for thornado and weaklib to allow reuse
These changes will allow existing libthornado.a and libweaklib.a
files to be reused, rather being rebuilt every time setup.py is
run (for a configuration that requires those libraries).
Currently this potential reuse will ONLY happend when setup
is invoked with the -noclobber flag.

Other parts of the setup system already internally cancel a -noclobber
requested from the command line when it is detected that LIBRARY
requests [see setup_libraries] or some proprocessor definitions
[see setup_defines] have changed from a previous setup invocation
(using the same object directory), so this extended effect of the
-noclobber flag should be fairly safe.

Caution is still advised when using -noclobber with (these) libraries.
While setup can detect some changes in the way setup is invoked
(including changes induced by switching from some simulations to
another), and can override -noclobber in those cases to be on the
safe side, it will (currently) **not** detect changes in the library
source code, such as might be caused by a change in repository commit
if the library code is retrieved from a git submodule.
Thus when -noclobber is used, updates of library code will **not**
trigger library recompilation and rebuild.