(Issue #215) Added in Milhoja Init unit test.  I have tried to structure this in accord with the unitTest architecture in the User Manual.  One main consequence of this is that it uses the generic Grid unitTest evolveAll routine, which writes a unitTest output file.  Since this test also uses the ut_testDriverMod framework, I had to simplify that so that it doesn't attempt to write the same file.  This change will likely affect the AMReX unit tests.
I have run this successfully in 1D, 2D, and 3D on GCE/compute-12 with Intel.
The test correctly creates a unitTest_XXXX output file and adds in a success
line only if the test was 100% successful.  I temporarily dumped the ICs and
final solution to AMReX-format files and manually confirmed correct content.