(Issue #215) Added in Milhoja Fluxes unit test following the structure of the Init test.
I have run this successfully in 1D, 2D, and 3D on GCE/compute-12 with Intel.
The test correctly creates a unitTest_XXXX output file and adds in a success
line only if the test was 100% successful.  I manually reviewed the milhoja.log
file to confirm that its output is consistent with expectations.  I reviewed the
logs for the Init test when it was added, but forgot to mention that in the
commit message.