(Issue #242) Fixed up Gravity unit test driver so that it no longer uses ut_testDriverMod.F90.  It was only using those facilities to create the PASS/FAIL unit test output file.  However, as part of Issue #215 those facilities no longer create this file.  Therefore, the driver now does this itself.
Ran the unit test on GCE using all default settings.  It now builds and runs
without error.  Confirmed that unitTest_0000 was created and that its contents
indicate success.  Added the line `potErrorMax = 1.0e300` to Gravity_unitTest
before that value is checked.  Confirmed in the unit test output that the test
failed.  Confirmed that unitTest_0000 was created and that its contents did not
include a line indicating success.