(Issue #215) Add in code for running Sedov with runtime on CPU.
Added in manually-generated code for running the Sedov problem with the CPU-only
TT config.  As part of this, the ICs can now be set with the runtime using the
CPUs as well.

With the runtime, we need Fortran code to be thread-safe.  When building with
Intel, we need to use the -auto flag.  Since this might be undesirable for
non-runtime builds, we now need an Intel and a Intel_Runtime version of the GCE
Makefiles.  Also, par files are now site specific so that we can set
Orchestration runtime parameters for the target platform.

Ran on GCE/compute012 the 2D/Pseudo-UG/Sedov/simpleUnsplit problem with Milhoja
but no runtime and built using the Intel makefile.  Reran with the runtime using
the Intel_Runtime makefile.  Both run through.  The no-runtime result is bitwise
exact to GCC-based results.  The runtime version is different at about
double-precision round-off error.  A difference is expected since the Hydro code
used by the runtime was ported from the C++ Sedov code in the Milhoja repo
and has a different structure.  This was done thinking toward a performance
study that compares C++ and Fortran runtime runs.  So far, so good.