(Issue #215) Continued cleaning & improving Milhoja/Flash-X coupling.
The Milhoja Grid and Orchestration Config files are improved so that these
configurations are done correctly based on setup-time settings.  As part of
this, users must now specify at setup a Milhoja setup variable in which they
specify backends and offloading.

Ran [23]D Mpi, CPU-only, and GPU-only tests on Summit.  All built and ran
through with an appropriate number of steps.  No other verification done.