(Issue #215) UnitTemplate doxygen docs not in API.
At reviewer Steve's request, I split the current Doxyfile into two, which was
already something that needed to be done.  One version is the user-based
API-only view intended for the public website; the other, for
developers/maintainers.  With this, the MyUnit.dox file can correctly be set as
non-@internal but not be rendered with the user-oriented API doxygen view.

Rendered both and confirmed that UnitTemplate not in API view.