New Sedov-specific TileWrapper class.
The class is effectively manually-generated since it is hardcoded to work with
both 2D and 3D for now.  Updated the TF so that it now gets dt from the data
item rather than from Driver.  Nearly there...

Ran the [23]D Sedov problems using the CPU-only orchestration setup.  Seemed to
work well.