tweaks for geeks
Add /lib/milhoja and a /lib/milhoja/Config that has a line

   LIBRARY amrex

To express that "milhoja depends on amrex".  Our setup command
will then "pull in" the AMReX library when the Milhoja library
is requested, and automatically order them correctly for proper
use on the linker ccommand line.

Add a fake

   LIB_OACC  =

to Makefile.h.Intel.testEnv, too.

Correct a missing hyphen in a setup command line for a Sedov test
with Milhoja.