Zero-initialize level-wide flux arrays for Spark...
using preexisting subroutine Grid_zeroFluxData,
which was slightly modified with some preprocessor tests
so that zero-filling only happens for the first NDIM directions.

The call to Grid_zeroFluxData was placed so it anly is done
for the per-level loop in Hydro.F90 since that is where a
problem was detected; but maybe it should be more general.

This zero-initialization at the beginnig of each Hydro step
may not be necessary:
* It could be replaced by putting zero-initialization of the
  level-wide flux arrays in the Amrex subroutines
  + gr_initNewLevelCallback.F90
  + gr_remakeLevelCallback.F90
  + gr_makeFineLevelFromCoarseCallback
  (because zeroing the data once at creation of the flux multifabs
  may be sufficient to avoid runtime errors).
* I suspect, based on debugging, that not initializing the level-
  wide flux multifab data only creates problems if the uninitialized
  data contains raw data that is interpreted as NaN (or maybe Inf).
  For in that situation a statement like
      new_value = wt * old_value + update ,
      with old_value containing NaN or Inf,
  can cause trouble EVEN WHEN wt has the value 0.0, so that the
  statement is then NOT equivalent to
      new_value = update  .
  Whether not explicitly initializing the level-wide flux data
  results in runtime errors or not can dpend on various compilation
  flags.

(cherry picked from commit 7c837ac4e00842ddecd8fa2df1b2d0c553c0c5dc)