Comment out EVOLVEDVAR in Sedov Config
Reordering of UNK variables, as via the recently included EVOLVEDBAR
directive, has caused significant differences in one Sedov test that
was recdently added to our regular testing, namely
  Composite_Sedov23_AMR_2dCyl_amrexLwf_spark_fcAllLev .

(There was no such effect in the corresponding Paramesh4 test
  Composite_Sedov23_AMR_2dCyl_pm4Lwf_spark_fcAllLev .)

It should be noted that the test
  Composite_Sedov23_AMR_2dCyl_amrexLwf_spark_fcAllLev
enables several features that are eiter not at all or very little
tested, especially in this combination:
 * Amrex Grid
 * enabled guard gell masking (enableMaskedGCFill = .TRUE.)
 * Spark Hydro
 * real AMR
 * initialization by mapping from analytical solution
 * use_hybridRiemann = .true.

This commit comments out the EVOLVEDVAR directive, so that the test
should revert to the previously apprived results.

The cause(s) of the result change induced by the EVOLVEDVAR sgould
be investigated separately. Depending on the outcome of the
investigstion, the EVOVEDVAR may eventually be enabled again, either
here or in some other location(s).