Disable diagnostic UNK variables tmph and tmpw
This commit is commenting out

   VARIABLE TMPH
   VARIABLE TMPW

and disabling associated logic for "comparing inverted temperatures"
between Helmholtz and Weaklib computation, which was specific to
the Hybrid EOS implementation and was relying on some hackery
that doesn't quite work any more after recent Eos reorganization.
This temperature comparison was only used for diagnostic output
anyway and was not part of the test criteria for success.