Edited Flame_rhJump.F90 to work with new Eos calls
Main changes: mask is obsolete; new interfaces Eos and Eos_vector;
Eos vector and derivative vectors are separate; and
these vectors are 2D instead of 1D when using the Eos_vector function.
To solve the latter problem we introduced a simple buffer to move the 1D data to 2D temporarily for the function call to Eos_vector.