Put in mesa-based species property data
Not all species used in the mesa networks are in the default
flash table of species data.  (needed for A and Z of each as
well as binding energies.)  This updates the supplemental python
script provided with the mesa Burn Unit to create this from one of
the data files in mesa, and also puts in the pieces to use it instead
of the usual table.
Note that the SpeciesList.txt file must be in the Burn tree in order
to overried the default file.  Not sure if that setup behavior should
be improved or if I'm just not using in quite right.