Make "Makefile" work for both older and new versions of GNU Make
...hopefully without warnings.

Depending on the Make version detected, used basically the old
mechanism for rules that involve dependencies on Fortran module or,
if the version is 4.3 or later, use a new way of writing rules that
use the '?:' separator.

This is a still a bit of experimental... and could be simplified.
(Currently the version dewtection is done in several ways / more
than one place.)