Fully ported Simulation_initBlock.F90 for Flame3StageNoise to Flash-X.
The template for this conversion was the Simulation_initBlock.F90 for Flame1StageNoise which was already converted to Flash-X by Catherine. For the Flame3 additions, much of the language that needed conversion was simple like 'state' -> 'eosData', but I needed to also add in the variables such as dyi_qn, dqbar_qn, that don't exist in Flame1StageNoise but exist in Flame3. I figured out how to use the 'deltas' by looking at the gradient detonation problem which already has them working in it's own Simulation_initBlock.F90 file.