Adding bn_paraInterface.F90 from Flash4 and need to port to Flash-X.
So there are quite a few files in the Flash4 parametric directory tree. At the first level of the directory tree, there is CONe2NSE (which I have added) as well as CONeNe2NSE which I have yet to add as the Config file (which I have added) stated that CONe2NSE was the default. In the CONe2NSE directory there are many different burn subroutines that I assume will all need to get ported to Flash-X. There also are other files like eos_externalComputeAbarZbar.F90 that I'm not sure about since that file exists elsewhere already ported to Flash-X. That is a dean question.