Adding the burn subroutines required for successful compilation.
Also changed a 'Flash.h' to 'Simulation.h'. So now, the Flame3StageNoise problem succesfully compiles but does not run for long as eos_helm.F90 gets a divide by zero error due to a zero value for abar. This means the next step is adding in the burn computeAbarZbar subroutine. A bit confused though because I don't think I made any changes to this routine when I got Flame1StageNoise to run. My biggest contribution was to eos_multiDim.F90 which simply called the computeAbarZbar routine. Not sure. Will have to think about this.